+++ get_x_navigation_generator_annotations/lib/get_x_navigation_generator_annotations.dart
@@ -1,0 +1,3 @@
+library get_x_navigation_generator_annotations;
+
+export 'src/annotations.dart';