+++ get_x_navigation_generator_interface/lib/get_x_navigation_generator_interface.dart
@@ -1,3 +1,0 @@
-library get_x_navigation_generator_interface;
-
-export 'src/annotations.dart';